refactor(attend): split main.rs into cmd/* modules (#51)
main.rs had grown to 2131 lines through honest feature accretion
across ADRs 113, 114, 119, 121, and 123. The file is already
structured as a command dispatcher — `fn main()` is a flat match
over subcommand strings that calls `cmd_*` helpers, each of which
is almost fully self-contained with its own colocated helpers.
The seams are visible in the function order.

This commit extracts all per-command code into a `cmd::` tree
mirroring the ways-cli convention and collects shared addressing
helpers in a new `util` module.

main.rs drops to 137 lines — mod declarations plus the argv
dispatch match. No behavior change.

New layout:

  src/main.rs             137  dispatch match only
  src/util.rs              58  signals_base, count_signals,
                               encode_project, own_session_id,
                               get_groups
  src/cmd.rs               19  pub mod declarations
  src/cmd/run.rs          515  cmd_run_with_catchup,
                               DisclosureGovernor,
                               ScheduledSensor, collect_snapshot
  src/cmd/inbox.rs        273  cmd_inbox, cmd_inbox_read,
                               ParsedSignal, parse_signal,
                               is_valid_signal_id
  src/cmd/send.rs         318  cmd_send, cmd_reply,
                               identify_sender, detect_terminal,
                               find_closest_peer
  src/cmd/tune.rs         342  cmd_tune, apply_engagement_tune,
                               replace_engagement_section,
                               extract_json_str, parse_iso8601,
                               days_from_civil
  src/cmd/cleanup.rs      219  cmd_cleanup, run_cleanup,
                               CleanupStats, parse_duration_arg
  src/cmd/focus.rs        118  cmd_focus_new
  src/cmd/config_cmd.rs    90  display_config
  src/cmd/status.rs        84  cmd_status
  src/cmd/peers.rs         57  cmd_peers
  src/cmd/scene.rs         42  cmd_scene, cmd_scenes
  src/cmd/permissions.rs   39  cmd_permissions_audit

Deviations from the issue's proposed layout:

- `get_groups` lives in `util` rather than `cmd::scene`. It's used
  by 5 command modules (inbox, peers, send, scene, status, focus),
  so hosting it in `util` next to `signals_base` / `own_session_id`
  — the same addressing-layer it composes — is the natural home.
  The issue's suggestion to put it in `cmd::scene` would have made
  every command reach into a sibling cmd module.
- `cmd::run` came in at 515 lines rather than the estimated ~425
  because the comments and whitespace from the original file
  preserved the existing annotations verbatim. 15 lines over the
  500-line review threshold; splitting DisclosureGovernor out
  would fight the single-lifetime invariant the governor has with
  the run loop. Left as one module.
- `CleanupStats` and `run_cleanup` are `pub(crate)` in `cmd::cleanup`
  because `cmd::run` calls `run_cleanup` for its in-loop auto-sweep.
  This is the one cross-cmd dependency; everything else reaches
  down into `util` or back up into the pre-existing sibling modules
  (`config`, `sensors`, `state`, `emit`, `groups`, `scenes`).
- `cmd/config_cmd.rs` is named that way (rather than `cmd/config.rs`)
  to avoid confusion with the top-level `config` module. Its only
  public item is `display_config`.

Refs #51.
